% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupProject.R
\name{PMS.test}
\alias{PMS.test}
\title{Methods for partially matched samples}
\usage{
PMS.test(x, y, alternative = "two.sided", method = "weighted.z",
  data = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values in tumor samples}

\item{y}{a (non-empty) numeric vector of data values in normal samples}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}

\item{method}{a number indicating which method to use for partially mathced samples, 1 is modified t-statistic of Kim, 2 is corrected Z-test of Looney and Jones, 3 is MLE based test of Ekbohm under homoscedasticity, 4 is MLE based test of Lin and Stivers under heteroscedasticity, 5 is weighted Z-test combination (default).}

\item{data}{an optional matrix or data frame containing the variables in the formula formula. By default the variables are taken from environment(formula).}

\item{conf.level}{confidence level of the interval}
}
\value{
test-statistic and p-value
}
\description{
This function allows you to use p-values pooling approach or modified test statistics approach to deal with partially matched samplles.
}
\examples{
x=c(0,1,2,3,NA,NA,NA,9)
y=c(3,9,NA,NA,4,5,NA,NA)
PMS.test(x=x,y=y,method="5",alternative="two.sided")
}
