% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMStest.R
\name{PMS.test}
\alias{PMS.test}
\title{Methods for partially matched samples}
\usage{
PMS.test(x, y, alternative = c("two.sided", "less", "greater"),
  method = c("weighted.z", "modified.t", "corrected.z", "MLE.homo",
  "MLE.hetero"), data = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values with missing values.}

\item{y}{a (non-empty) numeric vector of data values with different missing values of \code{x}.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{method}{method which is implemented for partially matched samples, \code{"weighted.z"} is weighted Z-test combination (default), \code{"modified.t"} is modified t-statistic of Kim et al, \code{"corrected.z"} is corrected Z-test of Looney and Jones, \code{"MLE.homo"} is MLE based test of Ekbohm under homoscedasticity, \code{"MLE.hetero"} is MLE based test of Lin and Stivers under heteroscedasticity.}

\item{data}{an optional matrix or data frame containing the variables in the formula \code{formula}. By default the variables are taken from \code{environment(formula)}.}
}
\value{
A list with class \code{"htest"} containing the following components:

\item{statistic}{the value of the test statistic. Either Z-statistic or t-statistic.}

\item{parameter}{the degrees of freedom for the t-statistic.}

\item{p.value}{the p-value for the test.}

\item{estimate}{the estimated difference in means.}

\item{null.value}{the specified hypothesized value of the mean difference.}

\item{alternative}{a character string describing the alternative hypothesis.}

\item{method}{a character string indicating what type of test was performed.}

\item{data.name}{a character string giving the names of the data.}
}
\description{
This function allows you to use weighted Z-test or modified test statistics approach to deal with partially matched samples.
}
\details{
\code{x} and \code{y} both must have at least two numeric values and two missing values. The missing values of \code{y} must be in different position with the missing values of \code{x}. If not, there is no need to use this function, you can simply clean your data and try regular \code{"t.test"}.

\code{alternative = "greater"} is the alternative that x has a larger mean than y.

Partially matched samples can be viewd as data genereated from two experiemntal designs: (i) n1 mathced 
pairs, and (ii) independent groups with n2 and n3 per group, where both designs intend to estimate the 
same parameter. 

Here, we implement modified test statistics approach or weighted Z-test for partially mateched samples.
According to the False Discovery Rate (FDR), False Non-discovery Rate (FNR), and receiver operating characteristic curve (ROC),
the results suggest that weighting by the square root of sample size (weighted Z-test) performs well in various scenarios.

This function does not calculate the confidence interval, because the confidence interval has no direct relationship with mean difference in all of the five methods.

See \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3717400/} for more details.
}
\examples{
## Example 1: generate random normal data
x = rnorm(100,0,1)
y = rnorm(100,0.5,1)
x.with.na = c(x[1:20],NA,x[21:50],NA,NA,NA,x[51:100])
y.with.na = c(NA,NA,y[1:80],NA,y[81:100],NA)
PMS.test(x=x.with.na,y=y.with.na,method="weighted.z",alternative="two.sided")

## Example 2: a dataset with missing values
library(datasets)
data<-matrix(data=presidents,30,4,byrow = FALSE)
data1<-data.frame(data)
x<-data1$X1
y<-data1$X4
PMS.test(x=x,y=y,method="modified.t",alternative="greater")
}
\author{
Weifan Han <\email{woailm1993@yahoo.com}>

Fan Zhang <\email{frank08081993@gmail.com}>

Peishan Han <\email{peishan.han@stonybrook.edu}>

Yudong Li <\email{yudong.li@stonybrook.edu}>

Elizabeth Pemberton <\email{pembertonei@gmail.com}>
}
